requirements.txt
flask
torch
torchvision
tqdm
opencv-python
Pillow
mediapipe
